add a file and layout for api routes.Implemented for lodge complaint.
