+++ src/Modules/ComplaintManagement/ComplaintModule.module.css
@@ -1,53 +1,76 @@
 .fusionText {
-    font-size: 1rem;
-    @media (min-width: mantine-breakpoint-md) {
-      font-size: 1.5rem;
-    }
+  font-size: 1rem;
+  @media (min-width: mantine-breakpoint-md) {
+    font-size: 1.5rem;
   }
-  
-  .fusionCaretCircleIcon {
-    font-size: 2rem;
-    @media (min-width: mantine-breakpoint-md) {
-      font-size: 1.5rem;
-    }
+}
+
+.fusionCaretCircleIcon {
+  font-size: 2rem;
+  @media (min-width: mantine-breakpoint-md) {
+    font-size: 1.5rem;
   }
-  
-  .fusionCaretIcon {
-    font-size: 1rem;
-    @media (min-width: mantine-breakpoint-md) {
-      font-size: 1.25rem;
-    }
+}
+
+.fusionCaretIcon {
+  font-size: 1rem;
+  @media (min-width: mantine-breakpoint-md) {
+    font-size: 1.25rem;
   }
-  
-  .fusionActiveRecentTab {
-    background-color: #15abff13;
-    color: #15abff;
-    font-weight: 800;
-    border-radius: 4px;
-  }
-  
-  .fusionTabsContainer {
-    flex-shrink: 1;
-    max-width: 80vw;
-    overflow-x: auto;
-    -ms-overflow-style: none;
-  }
-  
-  .fusionTabsContainer::-webkit-scrollbar {
-    display: none;
-  }
-  
-  .selectinputs {
-      background-color: #15abff20;
-      border-color: #15abff4d;
-  }
-  
-  .selectinputs::placeholder {
-      color: #15abff;
-      font-weight: 600;
-  }
-  
-  .selectoptions:hover {
-    background-color: #15abff10; /* Transparent blue hover background */
-  }
-  
+}
+
+.fusionActiveRecentTab {
+  background-color: #15abff13;
+  color: #15abff;
+  font-weight: 800;
+  border-radius: 4px;
+}
+
+.fusionTabsContainer {
+  flex-shrink: 1;
+  max-width: 80vw;
+  overflow-x: auto;
+  -ms-overflow-style: none;
+}
+
+/* Hide scrollbar for Webkit browsers */
+.fusionTabsContainer::-webkit-scrollbar {
+  display: none;
+}
+
+/* Scrollbar styling for other parts of the page */
+.scrollbar {
+  overflow-y: auto; /* Enable vertical scrolling */
+}
+
+.scrollbar::-webkit-scrollbar {
+  width: 8px; /* Width of the scrollbar */
+}
+
+.scrollbar::-webkit-scrollbar-track {
+  background: #f0f0f0; /* Light grey track color */
+  border-radius: 10px;
+}
+
+.scrollbar::-webkit-scrollbar-thumb {
+  background-color: #c0c0c0; /* Grey thumb color */
+  border-radius: 10px;
+}
+
+.scrollbar::-webkit-scrollbar-thumb:hover {
+  background-color: #a0a0a0; /* Darker grey on hover */
+}
+
+.selectinputs {
+  background-color: #15abff20;
+  border-color: #15abff4d;
+}
+
+.selectinputs::placeholder {
+  color: #15abff;
+  font-weight: 600;
+}
+
+.selectoptions:hover {
+  background-color: #15abff10; /* Transparent blue hover background */
+}
